refactor(functions): retire the RUJI merge/unmerge and Gaia SWITCH forms
The RUJI merge programme has concluded — no user can still open a mergeable
position — and SWITCH was a one-way migration entry point into THORChain that
has served its purpose. A form that can no longer be honoured is worse than no
form, so all three come out of the Functions surface:

- Thorchain/Merge/ (screen, view model, ThorchainMergeAsset) and
  TransactionBuilder/Merge/MergeTransactionBuilder.swift
- Thorchain/Unmerge/ (screen, view model, MergeTokenCatalog) and
  TransactionBuilder/Merge/{UnmergeShares,UnmergeTransactionBuilder}.swift
- Cosmos/Switch/ (screen, view model, SwitchAmount) and
  TransactionBuilder/Cosmos/SwitchTransactionBuilder.swift

with their wiring: the three FunctionTransactionType cases and the arms that
render them, the FunctionActionCatalog subtitle/icon entries, the legacy
FunctionCallType cases and their FunctionCallTypeMigration mappings, and the
now-orphaned merge catalog in Core/Models — whose file is renamed to
ThorchainContractConstants.swift for the staking contract addresses that are
all it still holds. Sixteen localization keys left with no caller are dropped
from all eight shipping locales.

The signing layer is deliberately untouched. thorchain.swift's .thorMerge /
.thorUnmerge wasm arms and CosmosHelperStruct's `switch:` memo branch read
VSTransactionType values that come from the shared VultisigCommonData
protobuf, and both Android (UnMergeStrategy.kt, SwitchStrategy.kt) and Windows
(UnmergeSpecific, ChainAction's 'switch') still ship these operations. In a
multi-device vault the proposing signer may be one of those platforms, so iOS
has to remain able to sign a payload it is handed even though it can no longer
propose one. Balance display is untouched for the same reason: BalanceService
still reads merge-account sizes so an existing merged position keeps rendering.

Gaia drops to a single operation and so takes the action list's single-action
passthrough instead of a one-row list; its getDefault now names that operation,
which retires the vestigial-default exemption it needed while SWITCH was
migrated but unremoved.
