+++ VultisigApp/VultisigApp/Features/FunctionTransaction/Common/Model/FunctionTransactionType.swift
@@ -85,36 +85,12 @@ enum FunctionTransactionType: Hashable {
     /// builds, and a bare `.vote` intent would invite a caller to route one
     /// through the other.
     case dydxVote(coin: CoinMeta)
-    /// Cosmos Hub → THORChain SWITCH. `coin` is the source chain's native
-    /// asset: the transfer goes to THORChain's inbound vault for that chain,
-    /// and that vault only credits the native asset — a Gaia IBC token sent
-    /// there is simply lost. The destination is deliberately absent from the
-    /// intent; it is the live inbound vault address, resolved when the
-    /// transaction is built rather than carried from wherever the caller came
-    /// from. The THORChain address the memo names comes from the vault, so it
-    /// is not on the intent either.
-    case theSwitch(coin: CoinMeta)
     /// THORChain secured-asset redemption (`SECURE-`). `coin` is THORChain's
     /// own native asset, deliberately *not* the asset being redeemed: which
     /// secured denoms a vault holds is a live bank-balance query, so no caller
     /// can name one upfront, and the native coin is the account those balances
     /// hang off. The form's picker resolves the redeemed coin from it.
     case withdrawSecuredAsset(coin: CoinMeta)
-    /// Rujira MERGE on THORChain. `coin` is the chain's native asset — the
-    /// anchor and the fee asset — not the coin the transaction is built
-    /// against: the form picks that from the merge catalog intersected with
-    /// the vault's holdings, so every coin it can reach is already held.
-    /// `denom` optionally names the catalog entry to open on, so a caller that
-    /// already knows which token the user means (a position card) can
-    /// pre-select it, exactly as `.leave(coin:node:)` pre-fills its address.
-    case merge(coin: CoinMeta, denom: String?)
-    /// THORChain RUJI UNMERGE — withdrawing merge shares back into the merged
-    /// token. `coin` is THORChain's native asset: the merged tokens live inside
-    /// the merge contract rather than the wallet, so the only coin this form
-    /// needs the vault to resolve is the one that pays the fee. `denom` opens
-    /// the picker on a token the caller already knows (`thor.kuji`, …); nil
-    /// leaves it on the first offered one.
-    case unmerge(coin: CoinMeta, denom: String?)
     /// THORChain node REBOND. `node` pre-fills the node currently holding the
     /// bond when the caller already knows it, mirroring `.leave(coin:node:)`;
     /// the memo's second address and the optional partial amount are always
@@ -174,16 +150,10 @@ enum FunctionTransactionType: Hashable {
             return [coin]
         case .dydxVote(let coin):
             return [coin]
-        case .theSwitch(let coin):
-            return [coin]
         case .withdrawSecuredAsset(let coin):
             // Only the native coin: the secured assets themselves are added to
             // the vault by the form as it discovers which ones the account
             // actually holds, so they cannot be pre-resolved here.
-            return [coin]
-        case .merge(let coin, _):
-            return [coin]
-        case .unmerge(let coin, _):
             return [coin]
         case .rebond(let coin, _):
             return [coin]
