+++ VultisigApp/VultisigApp/Features/FunctionTransaction/FunctionTransactionScreen.swift
@@ -285,39 +285,10 @@ struct FunctionTransactionScreen: View {
                         onVerify: onVerify
                     )
                 }
-            case .theSwitch(let coin):
-                resolvingCoin(coinMeta: coin) { coin in
-                    SwitchTransactionScreen(
-                        viewModel: SwitchTransactionViewModel(coin: coin, vault: vault),
-                        onVerify: onVerify
-                    )
-                }
             case .withdrawSecuredAsset(let coin):
                 resolvingCoin(coinMeta: coin) { coin in
                     SecuredWithdrawTransactionScreen(
                         viewModel: SecuredWithdrawTransactionViewModel(coin: coin, vault: vault),
-                        onVerify: onVerify
-                    )
-                }
-            case .merge(let coin, let denom):
-                resolvingCoin(coinMeta: coin) { coin in
-                    MergeTransactionScreen(
-                        viewModel: MergeTransactionViewModel(
-                            coin: coin,
-                            vault: vault,
-                            initialDenom: denom
-                        ),
-                        onVerify: onVerify
-                    )
-                }
-            case .unmerge(let coin, let denom):
-                resolvingCoin(coinMeta: coin) { coin in
-                    UnmergeTransactionScreen(
-                        viewModel: UnmergeTransactionViewModel(
-                            coin: coin,
-                            vault: vault,
-                            initialDenom: denom
-                        ),
                         onVerify: onVerify
                     )
                 }
