+++ VultisigApp/VultisigAppTests/FunctionCall/FunctionCallMigrationSeamTests.swift
@@ -255,66 +255,6 @@ final class FunctionCallMigrationSeamTests: XCTestCase {
         )
     }
 
-    // MARK: - Switch
-
-    private static func makeGaiaToken(_ ticker: String = "FUZN") -> Coin {
-        FunctionCallFixture.makeCoin(
-            .gaiaChain,
-            ticker: ticker,
-            decimals: 6,
-            isNative: false,
-            address: FunctionCallFixture.cosmosAddress
-        )
-    }
-
-    func testSwitchMapsToTheSwitchIntent() {
-        guard case .theSwitch(let mappedCoin)? = FunctionCallType.theSwitch.migratedTransactionType(
-            coin: FunctionCallFixture.makeATOM(),
-            nodeAddress: nil
-        ) else {
-            return XCTFail("Switch must map to the switch intent")
-        }
-
-        assertIsNativeAsset(mappedCoin, chain: .gaiaChain, ticker: "ATOM")
-    }
-
-    /// THORChain's GAIA inbound vault credits ATOM and nothing else. The legacy
-    /// screen used whatever coin was selected, so opening Functions from one of
-    /// Gaia's IBC tokens and picking Switch would have sent that token to the
-    /// vault with no way back.
-    func testSwitchIsPinnedToTheChainsNativeAssetNotTheSelectedCoin() {
-        guard case .theSwitch(let mappedCoin)? = FunctionCallType.theSwitch.migratedTransactionType(
-            coin: Self.makeGaiaToken(),
-            nodeAddress: nil
-        ) else {
-            return XCTFail("Switch must map to the switch intent")
-        }
-
-        assertIsNativeAsset(mappedCoin, chain: .gaiaChain, ticker: "ATOM")
-    }
-
-    /// Same fail-closed property LEAVE has: a vault that cannot resolve the
-    /// native asset lands on the shared "not in vault" error rather than
-    /// switching whatever token happened to be selected.
-    func testSwitchTargetsTheNativeAssetEvenWhenTheVaultDoesNotHoldIt() {
-        let token = Self.makeGaiaToken()
-        let vault = FunctionCallFixture.makeVault(coins: [token])
-        XCTAssertNil(vault.nativeCoin(for: .gaiaChain), "Fixture must not hold ATOM")
-
-        guard case .theSwitch(let mappedCoin)? = FunctionCallType.theSwitch.migratedTransactionType(
-            coin: token,
-            nodeAddress: nil
-        ) else {
-            return XCTFail("Switch must map to the switch intent")
-        }
-
-        assertIsNativeAsset(mappedCoin, chain: .gaiaChain, ticker: "ATOM")
-        XCTAssertFalse(
-            vault.coins.map { $0.toCoinMeta() }.contains(mappedCoin),
-            "The vault cannot resolve the intent's coin, so the shared error view is what the user sees"
-        )
-    }
-
     func testVoteIntentResolvesTheCoinItNeeds() {
         let coin = Self.makeDydx()
         let intent = FunctionTransactionType.dydxVote(coin: coin.toCoinMeta())
@@ -367,198 +307,10 @@ final class FunctionCallMigrationSeamTests: XCTestCase {
         )
     }
 
-    func testSwitchIntentResolvesTheCoinItNeeds() {
-        let coin = FunctionCallFixture.makeATOM()
-        let intent = FunctionTransactionType.theSwitch(coin: coin.toCoinMeta())
-        XCTAssertEqual(intent.coins, [coin.toCoinMeta()])
-    }
-
-    func testSwitchStaysSelectableOnGaia() {
-        XCTAssertTrue(FunctionCallType.getCases(for: FunctionCallFixture.makeATOM()).contains(.theSwitch))
-    }
-
-    /// Gaia's default falls back to `.custom`, which is unreachable in
-    /// practice: every row Gaia offers now routes through the action list to a
-    /// migrated screen, and `.custom` is not itself one of Gaia's operations
-    /// (see `testGaiaDefaultIsVestigialAndUnreachable` below).
-    func testGaiaDefaultIsVestigialAndUnreachable() {
-        let atom = FunctionCallFixture.makeATOM()
-        XCTAssertFalse(FunctionCallType.getCases(for: atom).contains(FunctionCallType.getDefault(for: atom)))
-    }
-
     func testWithdrawSecuredAssetIntentResolvesTheCoinItNeeds() {
         let coin = Self.makeRune()
         let intent = FunctionTransactionType.withdrawSecuredAsset(coin: coin.toCoinMeta())
         XCTAssertEqual(intent.coins, [coin.toCoinMeta()])
-    }
-
-    // MARK: - The mapping (merge)
-
-    /// MERGE spends a catalog token, but the intent names the chain anchor —
-    /// the form resolves the spend coin from the vault's holdings. Selecting
-    /// Merge from RUNE therefore carries RUNE and no pre-selection.
-    func testMergeMapsToTheMergeIntentOnTheNativeAsset() {
-        guard case .merge(let mappedCoin, let denom)? = FunctionCallType.merge.migratedTransactionType(
-            coin: Self.makeRune(),
-            nodeAddress: nil
-        ) else {
-            return XCTFail("Merge must map to the merge intent")
-        }
-
-        assertIsNativeAsset(mappedCoin, chain: .thorChain, ticker: "RUNE")
-        XCTAssertNil(denom, "RUNE is not a mergeable token, so there is nothing to pre-select")
-    }
-
-    /// The legacy sub-model's `preSelectToken()` matched `thor.<ticker>` of the
-    /// coin the screen was on. It never fired because `ensureRuneCoin()` ran
-    /// first; carrying the denom on the intent makes it real.
-    func testMergePreselectsTheDenomWhenTheSelectedCoinIsMergeable() {
-        let kuji = FunctionCallFixture.makeCoin(
-            .thorChain,
-            ticker: "KUJI",
-            decimals: 8,
-            isNative: false,
-            address: FunctionCallFixture.thorAddress
-        )
-
-        guard case .merge(let mappedCoin, let denom)? = FunctionCallType.merge.migratedTransactionType(
-            coin: kuji,
-            nodeAddress: nil
-        ) else {
-            return XCTFail("Merge must map to the merge intent")
-        }
-
-        assertIsNativeAsset(mappedCoin, chain: .thorChain, ticker: "RUNE")
-        XCTAssertEqual(denom, "thor.kuji")
-    }
-
-    /// A node address belongs to the node functions; MERGE must ignore it
-    /// rather than smuggle it into the form.
-    func testMergeIgnoresACarriedNodeAddress() {
-        guard case .merge(_, let denom)? = FunctionCallType.merge.migratedTransactionType(
-            coin: Self.makeRune(),
-            nodeAddress: Self.thorNode
-        ) else {
-            return XCTFail("Merge must map to the merge intent")
-        }
-
-        XCTAssertNil(denom)
-    }
-
-    /// The intent's `coins` is what `needsCoinAddition` / `addCoins` read. The
-    /// pickable merge tokens are vault holdings by construction, so the anchor
-    /// is the only coin to resolve.
-    func testMergeIntentResolvesTheCoinItNeeds() {
-        let coin = Self.makeRune()
-        let intent = FunctionTransactionType.merge(coin: coin.toCoinMeta(), denom: "thor.kuji")
-        XCTAssertEqual(intent.coins, [coin.toCoinMeta()])
-    }
-
-    // MARK: - The mapping (unmerge)
-
-    private static func makeMergeToken(_ ticker: String = "KUJI") -> Coin {
-        FunctionCallFixture.makeCoin(
-            .thorChain,
-            ticker: ticker,
-            decimals: 8,
-            isNative: false,
-            address: FunctionCallFixture.thorAddress
-        )
-    }
-
-    func testUnmergeMapsToTheUnmergeIntent() {
-        guard case .unmerge(let mappedCoin, let denom)? = FunctionCallType.unmerge.migratedTransactionType(
-            coin: Self.makeRune(),
-            nodeAddress: nil
-        ) else {
-            return XCTFail("Unmerge must map to the unmerge intent")
-        }
-
-        assertIsNativeAsset(mappedCoin, chain: .thorChain, ticker: "RUNE")
-        XCTAssertNil(denom, "RUNE is not a merge token, so the picker opens on its own first entry")
-    }
-
-    /// The legacy form pre-selected the merge token matching the coin the user
-    /// opened Functions from.
-    func testUnmergePreSelectsTheMergeTokenTheUserCameFrom() {
-        guard case .unmerge(_, let denom)? = FunctionCallType.unmerge.migratedTransactionType(
-            coin: Self.makeMergeToken(),
-            nodeAddress: nil
-        ) else {
-            return XCTFail("Unmerge must map to the unmerge intent")
-        }
-
-        XCTAssertEqual(denom, "thor.kuji")
-    }
-
-    func testUnmergeLeavesTheDenomUnsetForATokenThatCannotBeMerged() {
-        guard case .unmerge(_, let denom)? = FunctionCallType.unmerge.migratedTransactionType(
-            coin: FunctionCallFixture.makeTCY(),
-            nodeAddress: nil
-        ) else {
-            return XCTFail("Unmerge must map to the unmerge intent")
-        }
-
-        XCTAssertNil(denom, "TCY has no merge contract, so there is nothing to pre-select")
-    }
-
-    /// The unmerge wasm execute is addressed by contract and attaches no coins,
-    /// so the only coin the vault has to resolve is the one paying the fee.
-    func testUnmergeIsPinnedToRuneNotTheSelectedToken() {
-        guard case .unmerge(let mappedCoin, _)? = FunctionCallType.unmerge.migratedTransactionType(
-            coin: Self.makeMergeToken(),
-            nodeAddress: nil
-        ) else {
-            return XCTFail("Unmerge must map to the unmerge intent")
-        }
-
-        assertIsNativeAsset(mappedCoin, chain: .thorChain, ticker: "RUNE")
-    }
-
-    /// Same fail-closed property LEAVE has: a vault that cannot pay the
-    /// THORChain fee lands on the shared "not in vault" error instead of opening
-    /// a form whose Continue could never produce a signable transaction.
-    func testUnmergeTargetsRuneEvenWhenTheVaultDoesNotHoldIt() {
-        let kuji = Self.makeMergeToken()
-        let vault = FunctionCallFixture.makeVault(coins: [kuji])
-        XCTAssertNil(vault.nativeCoin(for: .thorChain), "Fixture must not hold RUNE")
-
-        guard case .unmerge(let mappedCoin, _)? = FunctionCallType.unmerge.migratedTransactionType(
-            coin: kuji,
-            nodeAddress: nil
-        ) else {
-            return XCTFail("Unmerge must map to the unmerge intent")
-        }
-
-        assertIsNativeAsset(mappedCoin, chain: .thorChain, ticker: "RUNE")
-        XCTAssertFalse(
-            vault.coins.map { $0.toCoinMeta() }.contains(mappedCoin),
-            "The vault cannot resolve the intent's coin, so the shared error view is what the user sees"
-        )
-    }
-
-    func testUnmergeIntentResolvesTheCoinItNeeds() {
-        let coin = Self.makeRune()
-        let intent = FunctionTransactionType.unmerge(coin: coin.toCoinMeta(), denom: nil)
-        XCTAssertEqual(intent.coins, [coin.toCoinMeta()])
-    }
-
-    func testUnmergeStaysSelectableOnThorchain() {
-        XCTAssertTrue(FunctionCallType.getCases(for: Self.makeRune()).contains(.unmerge))
-    }
-
-    /// Every token the picker offers has to be one the builder can address, or
-    /// the wasm execute would be sent to an empty contract address.
-    func testEveryOfferedMergeTokenHasAContractAndAnAsset() {
-        XCTAssertFalse(MergeTokenCatalog.tokens.isEmpty)
-        for token in MergeTokenCatalog.tokens {
-            XCTAssertNotNil(
-                MergeTokenCatalog.contractAddress(for: token.thorchainAsset),
-                "\(token.thorchainAsset) is offered but has no merge contract"
-            )
-            XCTAssertEqual(token.asset.chain, .thorChain)
-            XCTAssertEqual(token.asset.contractAddress.lowercased(), token.thorchainAsset.lowercased())
-        }
     }
 
     // MARK: - The mapping (rebond)
@@ -718,10 +470,6 @@ final class FunctionCallMigrationSeamTests: XCTestCase {
     /// the dropdown is not a cosmetic regression.
     func testWithdrawSecuredAssetStaysSelectableOnThorchain() {
         XCTAssertTrue(FunctionCallType.getCases(for: Self.makeRune()).contains(.withdrawSecuredAsset))
-    }
-
-    func testMergeStaysSelectableOnThorchain() {
-        XCTAssertTrue(FunctionCallType.getCases(for: Self.makeRune()).contains(.merge))
     }
 
     /// Whatever a chain defaults to has to be something the dropdown offers,
